+++ sample-web/src/wasmJsMain/kotlin/com/aditlal/sampleweb/Main.kt
@@ -1,11 +1,11 @@
 package com.aditlal.sampleweb
 
 import androidx.compose.ui.ExperimentalComposeUiApi
-import androidx.compose.ui.window.CanvasBasedWindow
+import androidx.compose.ui.window.ComposeViewport
 
 @OptIn(ExperimentalComposeUiApi::class)
 fun main() {
-    CanvasBasedWindow(canvasElementId = "ComposeTarget") {
+    ComposeViewport(viewportContainerId = "ComposeTarget") {
         DashboardApp()
     }
 }